refactor: 将金额类型修改为float64

+++ internal/account/model/asset.go
@@ -5,8 +5,8 @@ type Asset struct {
 	UserId     int     `db:"user_id" json:"user_id,omitempty"`
 	AssetType  *int    `db:"asset_type" json:"asset_type" binding:"required,numeric"`
 	AssetName  string  `db:"asset_name" json:"asset_name" binding:"required,max=30"`
-	Balance    float32 `db:"balance" json:"balance" binding:"required,numeric"`
-	Init       float32 `db:"init" json:"-"`
+	Balance    float64 `db:"balance" json:"balance" binding:"required,numeric"`
+	Init       float64 `db:"init" json:"-"`
 	Icon       string  `db:"icon" json:"icon" binding:"required"`
 	Del        int     `db:"del" json:"-"`
 	IsTotal    int     `db:"is_total" json:"is_total" binding:"required"`
